refactor: provider server API for VS copilot
