+++ packages/fx-core/src/component/generator/copilotPlugin/helper.ts
@@ -29,6 +29,7 @@ import fs from "fs-extra";
 import { getLocalizedString } from "../../../common/localizeUtils";
 import { EOL } from "os";
 import { SummaryConstant } from "../../configManager/constant";
+import { manifestUtils } from "../../driver/teamsApp/utils/ManifestUtils";
 
 const manifestFilePath = "/.well-known/ai-plugin.json";
 const componentName = "OpenAIPluginManifestHelper";
@@ -106,6 +107,8 @@ export async function listOperations(
   context: Context,
   manifest: OpenAIPluginManifest | undefined,
   apiSpecUrl: string | undefined,
+  includeExistingAPIs: boolean,
+  teamsManifestPath: string | undefined,
   shouldLogWarning = true
 ): Promise<Result<ApiOperation[], ErrorResult[]>> {
   if (manifest) {
@@ -131,7 +134,24 @@ export async function listOperations(
     return err(validationRes.errors);
   }
 
-  const operations = await specParser.list();
+  let operations = await specParser.list();
+
+  // Filter out exsiting APIs
+  if (!includeExistingAPIs) {
+    if (!teamsManifestPath) {
+      throw new Error();
+    }
+    const manifest = await manifestUtils._readAppManifest(teamsManifestPath);
+    if (manifest.isOk()) {
+      const existingOperationIds = manifestUtils.getOperationIds(manifest.value);
+      operations = operations.filter(
+        (operation: string) => !existingOperationIds.includes(operation)
+      );
+    } else {
+      throw manifest.error;
+    }
+  }
+
   const sortedOperations = sortOperations(operations);
   return ok(sortedOperations);
 }
