+++ packages/fx-core/src/question/other.ts
@@ -284,7 +284,7 @@ export function selectTeamsAppValidationMethodQuestionNode(): IQTreeNode {
 
 export function copilotPluginAddAPIQuestionNode(): IQTreeNode {
   return {
-    data: apiSpecLocationQuestion(),
+    data: apiSpecLocationQuestion(false),
     children: [
       {
         data: apiOperationQuestion(false),
